add default value for http_status in ClientException
ClientException is not a virtual base class, hence it can be
initialized and string formatted, but the required property
'http_status' is not set by defaut, which will cause AttributeError
if we cannot convert a dedicate exception from response.

This patch fixes it by adding default value 'N/A' for http_status,
and set it with response.status_code when it is a ClientException
instance.

Change-Id: Ib9920a1f21885917fe7bc4562f1b54602112a473
Closes-Bug: #1586994