fix warnings in documentation build
Replace the :option: role markup with simpler literal text markup so
that options that are not defined anywhere in the documentation do not
produce warnings.

Turn on the -W option in the docs environment in tox so that local
developer builds of the documentation fail if there is a warning.

Change-Id: I4a97b1efbfddc6137668c23461a882fa589d4999
Signed-off-by: Doug Hellmann <doug@doughellmann.com>