Add Python3 yoga unit tests
This is a patch to ensure unit testing is in place for all the of
the tested runtimes for yoga.

See also the PTI in governance [1].

[1]: https://governance.openstack.org/tc/reference/project-testing-interface.html

This patch is manually created, because the previous patch was not
merged before Xena release and release bot didn't create the automated
patch at that time.

Change-Id: I902f342800d6ea7dda3eff2a10f7c9dd54fa6151