Unblock the gate and update python testing
This combines two commits in one

* Change aggregation method to mean

Commit 6061286c3133a4b85c4a72160960ecf6a82d02f8 handles
ignoring gnocchi api error when metric is not created.
However there are functional test failures when the
aggregation method is set to last for alarm type
gnocchi_aggregation_by_resources_threshold,

This fix changes the aggregration method to mean for
any create/update of alarm type
gnocchi_aggregation_by_resources_threshold.

* Update python testing as per zed cycle testing runtime

In Zed cycle, we have dropped the python 3.6/3.7[1] testing
and its support. Add release notes and update the python
classifier for the same.

[1] https://governance.openstack.org/tc/reference/runtimes/zed.html

Co-Authored-By: Ghanshyam Mann <gmann@ghanshyammann.com>

Depends-On: https://review.opendev.org/c/openstack/aodh/+/843764
Change-Id: Ia11f25fa8c36c6051a44994f2545e2b64a665593
Closes-Bug: #1974682