Fix capping of oslo.db
The original intent was to cap sqlalchemy and oslo.db to keep
compatibility in stable/2023.1. Though this now disallow us to release
bug fixes of oslo.db in stable/2023.1. By changing the cap of oslo.db
to MINOR version (<12.4.0) will solve our issue, as oslo.db should not
break compatibility with a stable release, but we still can consume bug
fix releases of oslo.db.

Change-Id: If075db0f4b1eb6b1b8b88add299c0409d8925e7d